+++ Dockerfile
@@ -1,11 +1,8 @@
-FROM node:16
+FROM node:18
 WORKDIR /app
 COPY package*.json ./
-RUN npm install
-RUN npm install -g expo-cli
+RUN npm install --omit=dev
 RUN npm install -g ts-node
-RUN npm install @types/react@17.0.21
 COPY . .
 EXPOSE 3000
-EXPOSE 80
-CMD ["expo", "start", "-p", "80"]
+CMD ["ts-node", "api/server.ts"]
