+++ package.json
@@ -23,6 +23,7 @@
     "expo-device": "^5.2.1",
     "expo-notifications": "^0.18.1",
     "expo-status-bar": "~1.4.4",
+    "expo-updates": "~0.16.4",
     "express": "^4.18.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
@@ -38,8 +39,7 @@
     "react-native-web": "~0.18.10",
     "react-navigation-header-buttons": "^10.0.0",
     "socket.io": "^4.6.2",
-    "socket.io-react-hook": "^2.4.1",
-    "expo-updates": "~0.16.4"
+    "socket.io-react-hook": "^2.4.1"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
